refactor: stop spawning pgrep/ps anywhere; share one libproc enumerator
The permission-dialog auto-approver spawned /usr/bin/pgrep with
waitUntilExit on every 250ms monitor tick. The libproc-based enumeration
introduced for documentation discovery now lives in ProxyCore as
ProcessEnumeration and serves both call sites; the pgrep output parser
and its test go away with the subprocess.
